+++ example/lib/main.dart
@@ -13,24 +13,26 @@ class MainApp extends StatefulWidget {
 }
 
 class _MainAppState extends State<MainApp> {
+  @override
+  Widget build(BuildContext context) {
+    return const MaterialApp(home: Example2());
+  }
+}
+
+class Example2 extends StatefulWidget {
+  const Example2({super.key});
+
+  @override
+  State<Example2> createState() => _Example2State();
+}
+
+class _Example2State extends State<Example2> {
   final Graph<Map> graph = Graph<Map>(
-    boxSize: const Size(200, 100),
+    boxSize: const Size(150, 50),
     items: [
-      {"title": 'CEO', "id": '1', "to": null},
       {
-        "title": 'HR Manager: John',
-        "id": '2',
-        "to": '1',
-      },
-      {
-        "title": 'HR Officer: Jane',
-        "id": '3',
-        "to": '2',
-      },
-      {
-        "title": 'Project Manager: Zuher',
-        "id": '4',
-        "to": '1',
+        "id": '1',
+        "text": 'Main Block',
       },
     ],
     idProvider: (data) => data["id"],
@@ -39,37 +41,49 @@ class _MainAppState extends State<MainApp> {
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      home: Scaffold(
+    return Container(
+      decoration: BoxDecoration(
+        gradient: LinearGradient(
+          colors: [
+            Colors.purple.shade600,
+            Colors.red.shade300,
+          ],
+          begin: Alignment.bottomLeft,
+          end: Alignment.topRight,
+        ),
+      ),
+      child: Scaffold(
+        backgroundColor: Colors.transparent,
         body: Center(
           child: OrgChart<Map>(
             graph: graph,
-            // curve: Curves.easeOut,
-            // duration: 500,
             isDraggable: true,
+            onTap: (item) {
+              graph.addItem({
+                "id": graph.uniqueNodeId,
+                "text": '',
+                "to": item["id"],
+              });
+              setState(() {});
+            },
+            onDoubleTap: (item) async {
+              String? text = await getBlockText(context, item);
+              if (text != null) setState(() => item["text"] = text);
+            },
             builder: (details) {
-              return Card(
-                color: details.beingDragged
-                    ? Colors.blue
-                    : details.isOverlapped
-                        ? Colors.green
-                        : null,
-                elevation: 10,
-                child: Center(
-                  child: Column(
-                    mainAxisSize: MainAxisSize.min,
-                    children: [
-                      Text(details.item["title"]),
-                      ElevatedButton(
-                        onPressed: () {
-                          details.hideNodes(!details.nodesHidden);
-                        },
-                        child: Text(details.nodesHidden
-                            ? 'Press to Unhide'
-                            : 'Press to Hide'),
-                      )
-                    ],
+              return GestureDetector(
+                child: Container(
+                  decoration: BoxDecoration(
+                    border: Border.all(
+                      color: Colors.black,
+                      width: 2,
+                    ),
+                    color: details.isOverlapped
+                        ? Colors.grey.shade600
+                        : Colors.grey.shade400,
+                    borderRadius: BorderRadius.circular(10),
                   ),
+                  child: Center(child: Text(details.item["text"])),
                 ),
               );
             },
@@ -96,11 +110,45 @@ class _MainAppState extends State<MainApp> {
         ),
         floatingActionButton: FloatingActionButton(
             child: const Text('Reset Position'),
-            onPressed: () {
+            onPressed: () async {
               graph.calculatePosition();
               setState(() {});
             }),
       ),
     );
+  }
+
+  Future<String?> getBlockText(
+      BuildContext context, Map<dynamic, dynamic> item) async {
+    final String? text = await showDialog(
+      context: context,
+      builder: (context) {
+        String text = item["text"];
+        return AlertDialog(
+          title: const Text('Enter Text'),
+          content: TextFormField(
+            initialValue: item["text"],
+            onChanged: (value) {
+              text = value;
+            },
+          ),
+          actions: [
+            TextButton(
+              onPressed: () {
+                Navigator.of(context).pop();
+              },
+              child: const Text('Cancel'),
+            ),
+            TextButton(
+              onPressed: () {
+                Navigator.of(context).pop(text);
+              },
+              child: const Text('OK'),
+            ),
+          ],
+        );
+      },
+    );
+    return text;
   }
 }